# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedStock AI - Hospital Supply Chain Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /forecasts:
    get:
      operationId: listForecasts
      tags: [forecasts]
      summary: List demand forecasts
      parameters:
        - name: skuId
          in: query
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: horizon
          in: query
          schema:
            type: integer
            enum: [30, 60, 90]
      responses:
        "200":
          description: List of forecasts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Forecast"
...
  /forecasts/{skuId}:
    get:
      operationId: getSkuForecast
      tags: [forecasts]
      summary: Get forecast for a specific SKU
      parameters:
        - name: skuId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SKU forecast with all horizons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkuForecastDetail"
  /forecasts/{skuId}/override:
    post:
      operationId: overrideForecast
      tags: [forecasts]
      summary: Override a forecast with manual adjustment
      parameters:
        - name: skuId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForecastOverrideBody"
      responses:
        "200":
          description: Forecast overridden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Forecast"

  # Expiry & Wastage
  /expiry/items:
    get:
      operationId: listExpiryItems
      tags: [expiry]
      summary: List items nearing expiry
      parameters:
        - name: withinDays
          in: query
          schema:
            type: integer
            default: 90
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Items nearing expiry
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpiryItem"
...
  /expiry/mark-resolved:
    post:
      operationId: markExpiryResolved
      tags: [expiry]
      summary: Mark an expiry item as returned/donated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkExpiryResolvedBody"
      responses:
        "200":
          description: Expiry item resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpiryItem"
...
components:
  schemas:
...
    Forecast:
      type: object
      properties:
        id:
          type: integer
        skuId:
          type: integer
        skuName:
          type: string
        departmentId:
          type: integer
        horizon:
          type: integer
          enum: [30, 60, 90]
        predictedQuantity:
          type: number
        confidenceLow:
          type: number
        confidenceHigh:
          type: number
        isOverridden:
          type: boolean
        overrideValue:
          type: number
          nullable: true
        overrideReason:
          type: string
          nullable: true
        projectedStockoutDate:
          type: string
          format: date
          nullable: true
        generatedAt:
          type: string
          format: date-time
      required:
        - id
        - skuId
        - skuName
        - departmentId
        - horizon
        - predictedQuantity
        - confidenceLow
        - confidenceHigh
        - isOverridden
        - generatedAt
    SkuForecastDetail:
      type: object
      properties:
        skuId:
          type: integer
        skuName:
          type: string
        currentStock:
          type: number
        forecasts:
          type: array
          items:
            $ref: "#/components/schemas/Forecast"
        dailyPredictions:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              predicted:
                type: number
              low:
                type: number
              high:
                type: number
            required:
              - date
              - predicted
              - low
              - high
        modelAccuracy:
          type: number
          nullable: true
      required:
        - skuId
        - skuName
        - currentStock
        - forecasts
        - dailyPredictions
...
    ForecastOverrideBody:
      type: object
      properties:
        horizon:
          type: integer
          enum: [30, 60, 90]
        overrideValue:
          type: number
        reason:
          type: string
      required:
        - horizon
        - overrideValue
        - reason
    ExpiryItem:
      type: object
      properties:
        id:
          type: integer
        skuId:
          type: integer
        skuName:
          type: string
        batchNumber:
          type: string
        expiryDate:
          type: string
          format: date
        daysUntilExpiry:
          type: integer
        quantity:
          type: number
        unit:
          type: string
        unitCost:
          type: number
        estimatedLoss:
          type: number
        departmentId:
          type: integer
        departmentName:
          type: string
        supplier:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, resolved_returned, resolved_donated, expired]
        urgency:
          type: string
          enum: [critical, high, medium, low]
      required:
        - id
        - skuId
        - skuName
        - batchNumber
        - expiryDate
        - daysUntilExpiry
        - quantity
        - unit
        - unitCost
        - estimatedLoss
        - departmentId
        - departmentName
        - status
        - urgency
...
    MarkExpiryResolvedBody:
      type: object
      properties:
        expiryItemId:
          type: integer
        resolution:
          type: string
          enum: [returned, donated]
        notes:
          type: string
      required:
        - expiryItemId
        - resolution
...